change the organization of the manager package
        - a manager file with an interface to follow and a docker file
          that implements the manager interface using the docker SDK
