+++ internal/manager/docker.go
@@ -1,0 +1,164 @@
+package manager
+
+import (
+	"context"
+	"encoding/base64"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/client"
+	"github.com/docker/docker/pkg/archive"
+	"github.com/docker/docker/pkg/stdcopy"
+	"github.com/docker/go-connections/nat"
+	"io"
+	"io/ioutil"
+	"os"
+	"time"
+)
+
+type DockerManager struct {
+	cli  *client.Client
+	auth string
+}
+
+func New() *DockerManager {
+	cli, err := client.NewClientWithOpts(client.FromEnv)
+	if err != nil {
+		panic(err)
+	}
+	auth, err := generateAuth()
+	if err != nil {
+		panic(err)
+	}
+
+	return &DockerManager{cli, auth}
+}
+
+// Builds a Docker image
+func (d DockerManager) BuildImage(directory string, tag string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), time.Duration(300)*time.Second)
+	defer cancel()
+
+	// Get a tar file from directory
+	dockerfileTarReader, err := archive.TarWithOptions(directory, &archive.TarOptions{})
+	if err != nil {
+		return err
+	}
+
+	resp, err := d.cli.ImageBuild(
+		ctx,
+		dockerfileTarReader,
+		types.ImageBuildOptions{
+			Dockerfile: "Dockerfile",
+			Tags:       []string{tag},
+			NoCache:    true,
+			Remove:     true,
+		})
+	if err != nil {
+		fmt.Println(err, " :unable to build docker image")
+		return err
+	}
+
+	// block until the image is finished building
+	_, err = ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// Push an image to remote repository
+func (d DockerManager) PushImage(image string) error {
+	ctx := context.Background()
+	fmt.Println("Going to push " + image)
+	out, err := d.cli.ImagePush(ctx, image, types.ImagePushOptions{
+		RegistryAuth: d.auth,
+	})
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+	io.Copy(os.Stdout, out)
+	return nil
+}
+
+// Generate the dockerhub credentials from ENV variables
+func generateAuth() (string, error) {
+	username := os.Getenv("DOCKER_USERNAME")
+	password := os.Getenv("DOCKER_PASSWORD")
+
+	if username == "" || password == "" {
+		return "", errors.New("Missing Dockerhub username or password")
+	}
+
+	authConfig := types.AuthConfig{
+		Username: username,
+		Password: password,
+	}
+	encodedJSON, err := json.Marshal(authConfig)
+	if err != nil {
+		return "", err
+	}
+	authStr := base64.URLEncoding.EncodeToString(encodedJSON)
+	return authStr, nil
+}
+
+// Pull an image from a remote repository
+func (d DockerManager) PullImage(name string) error {
+	ctx := context.Background()
+	out, err := d.cli.ImagePull(ctx, name, types.ImagePullOptions{})
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+	io.Copy(os.Stdout, out)
+	return nil
+}
+
+// Create and start a container from a local image
+func (d *DockerManager) RunContainer(image string) error {
+	ctx := context.Background()
+
+	resp, err := d.cli.ContainerCreate(ctx, &container.Config{
+		Image: image,
+		ExposedPorts: nat.PortSet{
+			"8080/tcp": struct{}{},
+		},
+	}, nil, nil, "")
+	if err != nil {
+		return err
+	}
+
+	err = d.cli.ContainerStart(ctx, resp.ID, types.ContainerStartOptions{})
+	if err != nil {
+		return err
+	}
+
+	statusCh, errCh := d.cli.ContainerWait(ctx, resp.ID, container.WaitConditionNotRunning)
+	select {
+	case err := <-errCh:
+		if err != nil {
+			panic(err)
+		}
+	case <-statusCh:
+	}
+
+	out, err := d.cli.ContainerLogs(ctx, resp.ID, types.ContainerLogsOptions{ShowStdout: true})
+	if err != nil {
+		return err
+	}
+
+	stdcopy.StdCopy(os.Stdout, os.Stderr, out)
+	return nil
+}
+
+func (d DockerManager) ContainerIP(ctx context.Context, id string) (string, error) {
+	co, err := d.cli.ContainerInspect(ctx, id)
+	if err != nil {
+		return "", err
+	}
+
+	return co.NetworkSettings.IPAddress, nil
+}
